Added stringifiation for JsError (#76)
JsErrors from the play json parsing were being returned as JsResultExceptions,
which would then be sent as a string to the user, which is not a very
readable syntax error. This commit changes the behavior to use the syntax
shown on https://www.playframework.com/documentation/2.5.x/ScalaJson for
turning the JsError from parsing into a JSON string that can be used
as an error message.