Allow control of Play JSON formatting (#136)
This addresses #135 by restoring the ability to send in a custom JsValue
printer if the user wants. It continues the current default behavior
of using pretty-printed JSON. In addition to the ability explicitly
override the unparser, you can also do this by having an implicit
variable in scope of type `JsValue => String`.